Try to get the total number of blocks in index.ts
